# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server 
  - url: https://smsj-men-restapi.herokuapp.com/api/
    description: Heroku remote server
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token



  schemas:
    car:
      brand: string
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: number
        price:
          type: number
        type:
          type: string
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /car:
    get:
      summary: Retrieves a list of Products
      description: Retrieves a list of products as JSON objects.
      responses:
        200:
          description: A list of product JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/car"
    
    post:
      summary: Create a new Product
      description: Create a new Product
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/car"

      responses:
        201:
          description: Product created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/car"
  
    put:
      summary: Updates a specific Product
      description: Updates a specific Product based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/car"

      responses:
        201:
          description: Product updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/car"

    delete:
      summary: Deletes a specific Product
      description: Deletes a specific Product based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: Product deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/car"

  


  /car/{id}:
    get:
      summary: Specific Product
      description: Retrieves a specific Product based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A Product in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/car"
    

  /car/random:
      get:
        summary: Random car
        description: Retrieves a random car based on it id.
        responses:
          200:
            description: A car in the format of a JSON object.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/car"


  /car/random/supercar:
      get:
        summary: Random supercar
        description: Retrieves a Random supercar based on it id.
        parameters:
          - in: path
            name: id
            required: true
            description: MongoDB id
            schema:
              type: string
        responses:
          200:
            description: A supercar in the format of a JSON object.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/car/random/supercar"

  
  /car/random/suv:
      get:
        summary: Random SUV
        description: Retrieves a Random SUV based on it id.
        parameters:
          - in: path
            name: id
            required: true
            description: MongoDB id
            schema:
              type: string
        responses:
          200:
            description: A SUV in the format of a JSON object.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/car/random/suv"

  
  /car/random/hypercar:
      get:
        summary: Random hypercar
        description: Retrieves a Random hypercar based on it id.
        parameters:
          - in: path
            name: id
            required: true
            description: MongoDB id
            schema:
              type: string
        responses:
          200:
            description: A hypercar in the format of a JSON object.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/car/random"


    

    